requirements.txt
pynput
pygame-ce